requirements.txt
django == 3.0
djangorestframework
